pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "docx2txt>=0.9",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.12.0",
    "grandalf>=0.8",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "langgraph>=1.0.1",
    "matplotlib>=3.10.6",
    "openai-agents>=0.3.3",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "ragas>=0.3.5",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.0",
    "sounddevice>=0.5.3",
    "tiktoken>=0.11.0",
    "unstructured>=0.18.14",
    "wavio>=0.0.9",
]
